Python implementation of the Viterbi algorithm
The file viterbi_function.py contains the function that performs the Viterbi algorithm for decoding in a Hidden Markov Model (HMM).

sim_func.py contains an example of application to determine if a system operates in risky conditions. In particular, the focus is on the number of transitions from state 6 to state 5.